update shell script

+++ update.sh
@@ -1,9 +1,11 @@
-rm -rf 404.html assets categories detail index.html page tags
+DIRECTORY='../blog_by_vuepress/blog/.vuepress/dist/'
+if [ "`ls -A $DIRECTORY`" = "" ]; then
+    echo "$DIRECTORY is empty, stop upload!!!"
+else
+    rm -rf 404.html assets categories detail index.html page tags
+    mv ../blog_by_vuepress/blog/.vuepress/dist/* .
+    git add --all
+    git commit -m "Update"
+    git push
+fi
 
-mv ../blog_by_vuepress/blog/.vuepress/dist/* .
-
-git add --all
-
-git commit -m "Update"
-
-git push
